Removed expvar and moved code to stats.go
